+++ packages/components/src/templates/next/components/internal/SearchableTable/shared/index.ts
@@ -1,0 +1,4 @@
+export {
+  SearchableTableClient,
+  type SearchableTableClientProps,
+} from "./SearchableTableClient"
